chore: change keep_class_names, keep_fn_names, and changed tests use field name
