+++ crates/node_binding/binding.d.ts
@@ -747,8 +747,8 @@ export interface RawLibraryOptions {
 export interface RawMinification {
   passes: number
   dropConsole: boolean
-  keepClassnames: boolean
-  keepFnames: boolean
+  keepClassNames: boolean
+  keepFnNames: boolean
   comments: "all" | "some" | "false"
   asciiOnly: boolean
   pureFuncs: Array<string>
