--- packages/rspack/tests/configCases/builtins/minify-keep-classnames/entry.js
@@ -1,13 +1,0 @@
-class KeepClass {
-	constructor() {
-		this.name = "test-keep-class-names";
-	}
-
-	getName() {
-		return this.name;
-	}
-}
-
-const keepClass = new KeepClass();
-
-keepClass.getName();
