+++ packages/rspack/tests/configCases/builtins/minify-keep-classnames/index.js
@@ -1,0 +1,15 @@
+class KeepClass {
+	constructor() {
+		this.name = "test-keep-class-names";
+	}
+
+	getName() {
+		return this.name;
+	}
+}
+
+it("should keep class names", () => {
+	const name = KeepClass.name;
+	console.log("name", name);
+	expect(name).toBe("KeepClass");
+});
