--- packages/rspack/tests/configCases/builtins/minify-keep-fn-names/entry.js
@@ -1,14 +1,0 @@
-function main() {
-	return "test-keep-fn-names";
-}
-
-function callMain() {
-	const call = main();
-
-	console.log(call);
-	return call;
-}
-
-main();
-
-callMain();
