+++ packages/rspack/tests/configCases/builtins/minify-keep-fn-names/index.js
@@ -1,0 +1,8 @@
+function main() {
+	return "test-keep-fn-names";
+}
+
+it("should keep fn names", () => {
+	const name = main.name;
+	expect(name).toBe("main");
+});
