+++ packages/rspack/tests/configCases/builtins/minify-keep-fn-names/webpack.config.js
@@ -1,11 +1,7 @@
 module.exports = {
-	entry: {
-		entry: "./entry.js",
-		main: "./main.js"
-	},
 	builtins: {
 		minifyOptions: {
-			keepFnames: true
+			keepFnNames: true
 		}
 	},
 	optimization: {
